refactor: move recommendations code into a new dir
