+++ lms/djangoapps/learner_home/utils.py
@@ -1,9 +1,7 @@
 """API utils"""
 
 import logging
-import requests
 
-from django.conf import settings
 from django.contrib.auth import get_user_model
 from django.core.exceptions import MultipleObjectsReturned
 from rest_framework.exceptions import PermissionDenied, NotFound
@@ -54,26 +52,3 @@ def get_masquerade_user(request):
         )
         log.info(success_msg)
         return masquerade_user
-
-
-def get_personalized_course_recommendations(user_id):
-    """Get personalize recommendations from Amplitude."""
-    headers = {
-        "Authorization": f"Api-Key {settings.AMPLITUDE_API_KEY}",
-        "Content-Type": "application/json",
-    }
-    params = {
-        "user_id": user_id,
-        "get_recs": True,
-        "rec_id": settings.REC_ID,
-    }
-    response = requests.get(settings.AMPLITUDE_URL, params=params, headers=headers)
-    if response.status_code == 200:
-        response = response.json()
-        recommendations = response.get("userData", {}).get("recommendations", [])
-        if recommendations:
-            is_control = recommendations[0].get("is_control")
-            recommended_course_keys = recommendations[0].get("items")
-            return is_control, recommended_course_keys
-
-    return True, []
