+++ lms/djangoapps/learner_home/views.py
@@ -26,7 +26,6 @@ from common.djangoapps.student.helpers import (
     cert_info,
     user_has_passing_grade_in_course,
 )
-from common.djangoapps.student.models import CourseEnrollment
 from common.djangoapps.student.views.dashboard import (
     complete_course_mode_info,
     credit_statuses,
@@ -34,7 +33,6 @@ from common.djangoapps.student.views.dashboard import (
     get_filtered_course_entitlements,
     get_org_black_and_whitelist_for_site,
 )
-from common.djangoapps.track import segment
 from common.djangoapps.util.course import (
     get_encoded_course_sharing_utm_params,
     get_link_for_about_page,
@@ -48,17 +46,11 @@ from lms.djangoapps.commerce.utils import EcommerceService
 from lms.djangoapps.courseware.access import administrative_accesses_to_course_for_user
 from lms.djangoapps.courseware.access_utils import check_course_open_for_learner
 from lms.djangoapps.learner_home.serializers import (
-    CourseRecommendationSerializer,
     LearnerDashboardSerializer,
 )
 from lms.djangoapps.learner_home.utils import (
     get_masquerade_user,
-    get_personalized_course_recommendations,
 )
-from lms.djangoapps.learner_home.waffle import (
-    should_show_learner_home_amplitude_recommendations,
-)
-from openedx.core.djangoapps.catalog.utils import get_course_data
 from openedx.core.djangoapps.content.course_overviews.models import CourseOverview
 from openedx.core.djangoapps.programs.utils import ProgramProgressMeter
 from openedx.core.djangoapps.site_configuration import helpers as configuration_helpers
@@ -581,96 +573,3 @@ class InitializeView(APIView):  # pylint: disable=unused-argument
         response_data = serialize_learner_home_data(learner_dash_data, context)
 
         return Response(response_data)
-
-
-class CourseRecommendationApiView(APIView):
-    """
-    API to get personalized recommendations from Amplitude.
-
-    **Example Request**
-
-    GET /api/learner_home/recommendation/courses/
-    """
-
-    authentication_classes = (
-        JwtAuthentication,
-        SessionAuthenticationAllowInactiveUser,
-    )
-    permission_classes = (IsAuthenticated, NotJwtRestrictedApplication)
-
-    def get(self, request):
-        """
-        Retrieves course recommendations details.
-        """
-        if not should_show_learner_home_amplitude_recommendations():
-            return Response(status=404)
-
-        general_recommendations_response = Response(
-            CourseRecommendationSerializer(
-                {
-                    "courses": settings.GENERAL_RECOMMENDATIONS,
-                    "is_personalized_recommendation": False,
-                }
-            ).data,
-            status=200,
-        )
-
-        try:
-            user_id = request.user.id
-            is_control, course_keys = get_personalized_course_recommendations(user_id)
-        except Exception as ex:  # pylint: disable=broad-except
-            logger.warning(f"Cannot get recommendations from Amplitude: {ex}")
-            return general_recommendations_response
-
-        # Emits an event to track student dashboard page visits.
-        segment.track(
-            user_id,
-            "edx.bi.user.recommendations.viewed",
-            {
-                "is_personalized_recommendation": not is_control,
-            },
-        )
-
-        if is_control or not course_keys:
-            return general_recommendations_response
-
-        recommended_courses = []
-        user_enrolled_course_keys = set()
-        fields = ["title", "owners", "marketing_url"]
-
-        course_enrollments = CourseEnrollment.enrollments_for_user(request.user)
-        for course_enrollment in course_enrollments:
-            course_key = f"{course_enrollment.course_id.org}+{course_enrollment.course_id.course}"
-            user_enrolled_course_keys.add(course_key)
-
-        # Pick 5 course keys, excluding the user's already enrolled course(s).
-        enrollable_course_keys = list(
-            set(course_keys).difference(user_enrolled_course_keys)
-        )[:5]
-        for course_id in enrollable_course_keys:
-            course_data = get_course_data(course_id, fields)
-            if course_data:
-                recommended_courses.append(
-                    {
-                        "course_key": course_id,
-                        "title": course_data["title"],
-                        "logo_image_url": course_data["owners"][0]["logo_image_url"],
-                        "marketing_url": course_data.get("marketing_url"),
-                    }
-                )
-
-        # If no courses are left after filtering already enrolled courses from
-        # the list of amplitude recommendations, show general recommendations
-        # to the user.
-        if not recommended_courses:
-            return general_recommendations_response
-
-        return Response(
-            CourseRecommendationSerializer(
-                {
-                    "courses": recommended_courses,
-                    "is_personalized_recommendation": not is_control,
-                }
-            ).data,
-            status=200,
-        )
